requirements.txt
cmudict
cn2an
faster-whisper==0.10.1
g2p_en
GPUtil
gradio>=4.32
jieba
librosa==0.9.2
loguru
num2words
numpy<2
protobuf==4.25
psutil
punctuators
pyannote.audio>=3.1.0
pyloudnorm
pyopenjtalk-dict
pypinyin
pyworld-prebuilt
stable_ts
tensorboard
torch
torchaudio
transformers
umap-learn
python-dotenv
load_dotenv
simpleaudio
sounddevice
keyboard
python-mpv
pygetwindow

